pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mas-mission-planning"
version = "0.1.0"
description = "DiCWO: Distributed Calibration-Weighted Orchestration for Satellite Mission Design"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30",
    "tiktoken>=0.7",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "jupyter>=1.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
docs = ["mkdocs-material>=9.5"]

[project.scripts]
run-experiment = "scripts.run_experiment:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]
